fixed an include issue that caused compilation to fail

+++ src/GameTime.h
@@ -21,6 +21,7 @@
 #define GAMETIME_H_
 #include <thread>
 #include <chrono>
+#include <GL/glew.h>
 #include <GL/glfw3.h>
 namespace MINX
 {
@@ -46,7 +47,7 @@ namespace MINX
 			void update();
 			/** limits the updates per second of the current thread by delaying
 			 */
-			inline void limitFPS(int desiredFPS){std::this_thread::sleep_for(std::chrono::milliseconds(1000/desiredFPS-(long(glfwGetTime())-lastUpdate)));}
+			inline void limitFPS(int desiredFPS){std::this_thread::sleep_for(std::chrono::milliseconds(1000/desiredFPS-(long(glfwGetTime())-long(lastUpdate))));}
 		private:
 			double totalTimeMillis;
 			double deltaTime;
